% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localSCR.R
\name{customize_model}
\alias{customize_model}
\title{Function to efficiently edit rows of model code generated from nimble}
\usage{
customize_model(
  model,
  append_code = NULL,
  line_append = NULL,
  line_remove = NULL,
  write = FALSE
)
}
\arguments{
\item{model}{\code{nimbleCode()} used to define model in \code{nimble} package,
possibly generated from \code{\link{get_classic}} function.}

\item{append_code}{either \code{NULL} or model code produced from 
\code{nimbleCode()}  or \code{\link{get_classic}} function. Note that if
\code{line_remove = NULL}, then code will be appended just after existing model
code; otherwise specify the lines to insert new code into by setting 
\code{line_append}.}

\item{line_append}{either \code{NULL} or an integer value as a scalar or vector
defining which positions to insert new lines of code in the model. Note that
line removal will occur first if \code{line_remove} is set, and then new
lines of code will be inserted at the desired indexing positions.}

\item{line_remove}{either \code{NULL} or an integer value as a scalar or vector
defining which lines of code to remove from model. Set to \code{NULL} when only 
appending to and not replacing code in previous model file.}

\item{write}{logical. If \code{TRUE}, then a text file is written to the 
working directory called "new_model.txt". Otherwise, model is written to temp
file and then deleted. Default is \code{FALSE}.}
}
\value{
a model description that can be run in \code{nimble} or using 
\code{\link{run_classic}}.
}
\description{
Allows for efficient editing of model code produced by \code{nimbleCode()} 
function
}
\examples{
# get model
scr_model = get_classic(dim_y = 2, enc_dist = "binomial",sex_sigma = TRUE,
hab_mask=TRUE,trapsClustered = TRUE)

# create new nimbleCode to use for replacement in 'scr_model'
p0_prior = nimble::nimbleCode({
   p0[g] ~ dbeta(1,1)
})

# replace line 3 of old model code with 'p0_prior' 
new_model = customize_model(model = scr_model, append_code = p0_prior, 
                                   line_append = 3, line_remove = 3)
                                   
# inspect new model code
new_model
}
\author{
Daniel Eacker
}
