% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localSCR.R
\name{discretize_classic}
\alias{discretize_classic}
\title{Function to discretize traps and initial activity centers}
\usage{
discretize_classic(X, grid, s.st = s.st, crs_, site, hab_mask = NULL)
}
\arguments{
\item{X}{Either a matrix or array representing the coordinates of traps in
UTMs. An array is used when traps are clustered over a survey area.}

\item{grid}{A matrix or array object of the the state-space grid. This is 
returned from \code{\link{grid_classic}}.}

\item{s.st}{A matrix of starting activity center coordinates. This is 
returned from \code{\link{initialize_classic}}}

\item{crs_}{The UTM coordinate reference system (EPSG code) used for your
location provided as an integer (e.g., 32608 for WGS 84/UTM Zone 8N).}

\item{site}{Either \code{NULL} (if a 2D trap array is used) or a vector of 
integers denoting which trap array an individual (either detected or 
augmented) belongs to. Note that \code{site} is provided from 
\code{\link{sim_classic}} when a 3D trap array is used. However, this
 \code{site} variable must be correctly augmented based on the total 
 augmented population size (i.e., \code{M}).}

\item{hab_mask}{Either \code{NULL} (the default) or a matrix or array output
from \code{\link{mask_polygon}} or \code{\link{mask_raster}} functions.}
}
\value{
A list of discretized traps as a matrix or array \code{X} and indices
for initial activity center locations. Note that the latter relates
to the rows of \code{grid} provided as an input.
}
\description{
Convert traps and initial activity center locations to 
prepare for using in discrete SCR model.
}
\details{
This function prepares the trap coordinates and initial activity
center coordinates for use in a discrete spatial capture-recapture model. Note
that the number of rows in the state-space grid coordinate matrix will be 
reduced and this object will need to be adjusted before being used in the model.
}
\examples{
# simulate a single trap array with random positional noise
x <- seq(-800, 800, length.out = 5)
y <- seq(-800, 800, length.out = 5)
traps <- as.matrix(expand.grid(x = x, y = y))
set.seed(200)
traps <- traps + runif(prod(dim(traps)),-20,20) 

mysigma = 300 # simulate sigma of 300 m
mycrs = 32608 # EPSG for WGS 84 / UTM zone 8N

# create state-space
Grid = grid_classic(X = traps, crs_ = mycrs, buff = 3*mysigma, res = 100)

# simulate data for uniform state-space and habitat mask
data3d = sim_classic(X = traps, ext = Grid$ext, crs_ = mycrs, sigma_ = mysigma,
prop_sex = 0.7,  N = 200, K = 4, base_encounter = 0.15, enc_dist = "binomial",
hab_mask = NULL, setSeed = 100)

# total augmented population size 
M = 400

# get initial activity center starting values
s.st = initialize_classic(y=data3d$y, M=M, X=traps, buff = 3*max(mysigma), 
                            hab_mask = NULL)$s.st

# get discretized traps and initial activity center grid indices
discrete_list <- discretize_classic(X = traps, grid=Grid$grid, 
                     s.st = s.st, crs_= mycrs,site=site,hab_mask=NULL)

str(discretize_list)
}
\author{
Daniel Eacker
}
