% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localSCR.R
\name{get_classic}
\alias{get_classic}
\title{Function to retrieve nimbleCode for spatially-explicit mark recapture models}
\usage{
get_classic(
  dim_y,
  enc_dist = "binomial",
  sex_sigma = FALSE,
  hab_mask = FALSE,
  trapsClustered = FALSE
)
}
\arguments{
\item{dim_y}{an integer of either 2 (the default) or that defines what 
dimensional format the encounter history data are in.}

\item{enc_dist}{either \code{"binomial"} or \code{"poisson"}. Default is
\code{"binomial"}.}

\item{sex_sigma}{a logical value indicating whether the scaling parameter 
('sigma') is sex-specific}

\item{hab_mask}{a logical value indicating whether a habitat mask will be 
used. Default is \code{FALSE}.}

\item{trapsClustered}{a logical value indicating if traps are clustered in 
arrays across the sampling area.}
}
\value{
a \code{nimbleCode} object from the \code{nimble} package.
}
\description{
Creates a \code{nimbleCode} object from the \code{nimble} package.
}
\details{
This function provides templates that could be copied and easily 
modified to include further model complexity such as covariates explaining 
detection probability. The models include different encounter probability 
distributions, sex-specific scaling parameters, and habitat masking.
}
\examples{
# get model for 2D encounter data, binomial encounter distribution, 
# non-sex-specific scaling parameter, and no habitat mask
scr_model = get_classic(dim_y = 2,enc_dist = "binomial",sex_sigma = FALSE,
                         hab_mask = FALSE,trapsClustered = FALSE)

# inspect model
scr_model
}
\author{
Daniel Eacker
}
