% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localSCR.R
\name{get_discrete}
\alias{get_discrete}
\title{Function to retrieve nimbleCode for discrete spatial capture-recapture models}
\usage{
get_discrete(
  dim_y,
  enc_dist = "binomial",
  sex_sigma = FALSE,
  trapsClustered = FALSE
)
}
\arguments{
\item{dim_y}{An integer of either 2 (the default) or that defines what 
dimensional format the encounter history data are in.}

\item{enc_dist}{Either \code{"binomial"} or \code{"poisson"}. Default is
\code{"binomial"}.}

\item{sex_sigma}{A logical value indicating whether the scaling parameter 
('sigma') is sex-specific}

\item{trapsClustered}{A logical value indicating if traps are clustered in 
arrays across the sampling area.}
}
\value{
A \code{nimbleCode} object from the \code{nimble} package.
}
\description{
Creates model code using the \code{\link[nimble]{nimbleCode}} function.
}
\details{
This function provides templates that could be copied and easily 
modified to include further model complexity such as covariates explaining 
detection probability. These discrete models include different encounter 
probability distributions and sex-specific scaling parameters for marked
data sets.
}
\examples{
# get discrete model for 2D encounter data, binomial encounter distribution, 
# non-sex-specific scaling parameter, and no clustering of traps
discrete_model = get_classic(dim_y = 2,enc_dist = "binomial",
                         sex_sigma = FALSE, trapsClustered = FALSE)

# inspect model
discrete_model
}
\author{
Daniel Eacker
}
