% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localSCR.R
\name{get_unmarked}
\alias{get_unmarked}
\title{Retrieve nimbleCode for spatial count models}
\usage{
get_unmarked(occ_specific = FALSE, hab_mask = FALSE, trapsClustered = FALSE)
}
\arguments{
\item{occ_specific}{Logical. If \code{FALSE}, the encounter rate will
not include an occasion-specific loop in the detection function; otherwise, 
the model will include a for loop for occasions (K) in the detection function.
Default is \code{FALSE}.}

\item{hab_mask}{A logical value indicating whether a habitat mask will be 
used. Default is \code{FALSE}.}

\item{trapsClustered}{A logical value indicating if traps are clustered in 
arrays across the sampling area.}
}
\value{
Model code created from \code{\link[nimble]{nimbleCode}}.
}
\description{
Creates model code using the \code{\link[nimble]{nimbleCode}} function.
}
\details{
This function provides templates for unmarked models that can be 
easily modified to include further model complexity such as covariates 
explaining detection probability. The models include habitat masking.
}
\examples{
# get spatial count model with non-occasion-specific detection
# function, single scaling parameter, no habitat mask, and no clustering
unmarked_model = get_unmarked(occ_specific=FALSE,hab_mask = FALSE, 
                              trapsClustered = FALSE)

# inspect model
unmarked_model
}
\author{
Daniel Eacker
}
