% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localSCR.R
\name{print_model}
\alias{print_model}
\title{Function to display row indices for model used in 'nimble'}
\usage{
print_model(model)
}
\arguments{
\item{model}{The \code{nimbleCode()} used to define model in \code{nimble} package,
possibly generated from \code{\link{get_classic}} function.}
}
\value{
A vector of characters with each row representing a line of the model. This
function is meant to help with using \code{\link{customize_model}}.
}
\description{
Prints model created with \code{nimbleCode()}to show row indices.
}
\examples{
# get model
scr_model = get_classic(dim_y = 2, enc_dist = "binomial",sex_sigma = TRUE,
hab_mask=TRUE,trapsClustered = TRUE)

print_model(scr_model)
}
\author{
Daniel Eacker
}
