% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localSCR.R
\name{rescale_classic}
\alias{rescale_classic}
\title{Function to rescale trap coordinates, grid extent, and starting activity 
center coordinates}
\usage{
rescale_classic(X, ext, s.st, site, hab_mask)
}
\arguments{
\item{X}{Either a matrix or array representing the coordinates of traps in
UTMs. An array is used when traps are clustered over a survey area.}

\item{ext}{An \code{Extent} object from the \code{raster} package. This is 
returned from \code{\link{grid_classic}}.}

\item{s.st}{A matrix of starting activity center coordinates. This is 
returned from \code{\link{initialize_classic}}.}

\item{site}{Either \code{NULL} (if a 2D trap array is used) or a vector of 
integers denoting which trap array an individual (either detected or 
augmented) belongs to. Note that \code{site} is provided from 
\code{\link{sim_classic}} when a 3D trap array is used. However, this 
\code{site} variable must be correctly augmented based on the total 
augmented population size (i.e., \code{M}).}

\item{hab_mask}{A matrix or arary output from \code{\link{mask_polygon}} or
\code{\link{mask_raster}} functions.}
}
\value{
A list of rescaled trap coordinates, grid extents, and starting 
activity center coordinates.
}
\description{
Rescale inputs to prepare data for habitat mask to be used.
}
\details{
This function is only meant to be used when habitat masking is 
incorporated into the model. The functions properly rescales inputs based on 
the dimensions of the habitat mask. Note that the \code{pixelWidth} needs to 
be included as an input in the model after inputs are rescaled to correctly
 estimate the scaling parameter (i.e., 'sigma').
}
\examples{
# simulate a single trap array with random positional noise
x <- seq(-800, 800, length.out = 5)
y <- seq(-800, 800, length.out = 5)
traps <- as.matrix(expand.grid(x = x, y = y))

# add some random noise to locations
traps <- traps + runif(prod(dim(traps)),-20,20) 

mysigma = 300 # simulate sigma of 300 m
mycrs = 32608 # EPSG for WGS 84 / UTM zone 8N

# create grid
Grid = grid_classic(X = traps, crs_ = mycrs, buff = 3*mysigma, res = 100)

# create polygon to use as a mask
library(sf)
poly = st_sfc(st_polygon(x=list(matrix(c(-1765,-1765,1730,-1650,1600,1650,
0,1350,-800,1700,-1850,1000,-1765,-1765),ncol=2, byrow=TRUE))), crs =  mycrs)

# make simple plot
par(mfrow=c(1,1))
plot(Grid$grid, pch=20)
points(traps, col="blue",pch=20)
plot(poly, add=TRUE)

# create habitat mask
hab_mask = mask_polygon(poly = poly, grid = Grid$grid, crs_ = mycrs, 
prev_mask = NULL)

# simulate SCR data
data3d = sim_classic(X = traps, ext = Grid$ext, crs_ = mycrs, 
sigma_ = mysigma, prop_sex = 1, N = 200, K = 4, base_encounter = 0.25, 
enc_dist = "binomial", hab_mask = hab_mask, setSeed = 50)

# generate initial activity center coordinates for 2D trap array without 
#habitat mask
s.st3d = initialize_classic(y=data3d$y, M=500, X=traps, ext=Grid$ext, 
hab_mask = hab_mask)

# build rescaled constants list for 2D trap array
constList = rescale_classic(X = traps, ext = Grid$ext, s.st = s.st3d, 
site = NULL, hab_mask = hab_mask)
str(constList)
}
\seealso{
\code{\link{mask_polygon}}, \code{\link{mask_raster}}
}
\author{
Daniel Eacker
}
