% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localSCR.R
\name{update_model}
\alias{update_model}
\title{Function to efficiently edit rows of model code generated from nimble}
\usage{
update_model(model, line_remove = NULL, append_code = NULL, write = FALSE)
}
\arguments{
\item{model}{\code{nimbleCode()} used to define model in \code{nimble} package,
possibly generated from \code{\link{get_classic}} function.}

\item{line_remove}{either \code{NULL} or a integer value defining which lines
of code to remove from model. Set to \code{NULL} when only appending to and
not replacing code in previous model file.}

\item{append_code}{either \code{NULL} or model code produced from 
\code{nimbleCode()}  or \code{\link{get_classic}} function. Note that if
\code{line_remove = NULL}, then code will be appended just after existing
model code; otherwise specify the lines to replace when appending by setting
\code{line_remove}.}

\item{write}{logical. If \code{TRUE}, then a text file is writing to the 
working directory called "new_model.txt". Otherwise, model is written to temp
file and then deleted. Default is \code{FALSE}.}
}
\value{
a model description that can be run in \code{nimble} or using 
\code{\link{run_classic}}.
}
\description{
Allows for efficient editing of model code produced by \code{nimbleCode()} function
}
\examples{
# get model
scr_model = get_classic(dim_y = 2, enc_dist = "binomial",sex_sigma = TRUE,
hab_mask=TRUE,trapsClustered = TRUE)

# create new nimbleCode to use for replacement in 'scr_model'
p0_prior = nimble::nimbleCode({
   p0[g] ~ dbeta(1,1)
})

# replace line 3 of old model code with 'p0_prior' 
new_model = update_model(model = scr_model, line_remove = 3, 
                         append_code = p0_prior)

# inspect new model code
new_model
}
\author{
Daniel Eacker
}
