% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localSCR.R
\name{get_discrete}
\alias{get_discrete}
\title{Function to retrieve nimbleCode for discrete spatial capture-recapture models}
\usage{
get_discrete(
  type = "marked",
  dim_y,
  occ_specific = FALSE,
  enc_dist = "binomial",
  sex_sigma = FALSE,
  trapsClustered = FALSE
)
}
\arguments{
\item{type}{Specifies the type of discrete model for either \code{"marked"}
(the default) or \code{"unmarked"} data sources.}

\item{dim_y}{An integer of either 2 (the default) or that defines what 
dimensional format the encounter history data are in. Note that this input
is not used when \code{type = "unmarked"}.}

\item{occ_specific}{Logical. If \code{FALSE}, the encounter rate will
not include an occasion-specific loop in the detection function; otherwise, 
the model will include a for loop for occasions (K) in the detection function.
Default is \code{FALSE}. Only applied when \code{type = "unmarked"}.}

\item{enc_dist}{Either \code{"binomial"} or \code{"poisson"}. Default is
\code{"binomial"}.}

\item{sex_sigma}{A logical value indicating whether the scaling parameter 
('sigma') is sex-specific}

\item{trapsClustered}{A logical value indicating if traps are clustered in 
arrays across the sampling area.}
}
\value{
A \code{nimbleCode} object from the \code{nimble} package.
}
\description{
Creates model code using the \code{\link[nimble]{nimbleCode}} 
function.
}
\details{
This function provides templates that could be copied and easily 
modified to include further model complexity such as covariates explaining 
detection probability. These discrete models include different encounter 
probability distributions and sex-specific scaling parameters for marked and
unmarked data sets.
}
\examples{
# get discrete model for 2D marked encounter data, binomial encounter 
# distribution, non-sex-specific scaling parameter, and no clustering of traps
discrete_model_m = get_discrete(type="marked", dim_y = 2,enc_dist = "binomial",
                      sex_sigma = FALSE, trapsClustered = FALSE)

# inspect model
discrete_model_m

# get discrete model for unmarked encounter data, binomial encounter 
# distribution, non-sex-specific scaling parameter, and no clustering of traps
discrete_model_u = get_discrete(type="unmarked",occ_specific = FALSE,
           enc_dist = "binomial",sex_sigma = FALSE, trapsClustered = FALSE)

# inspect model
discrete_model_u
}
\author{
Daniel Eacker
}
